backend/requirements.txt
finvizfinance
pandas
google-cloud-storage
google-cloud-bigquery
functions-framework
pyarrow
requests
